refactor: BuisinessException 발생 시 디버깅 용 오류 로깅 출력 기능 추가
