+++ src/main/java/gift/common/dto/ErrorResponse.java
@@ -1,10 +1,14 @@
-package gift.common;
+package gift.common.dto;
 
 import gift.common.exception.ErrorCode;
 
 public record ErrorResponse(String errorCode, String errorMessage) {
 
   public static ErrorResponse from(ErrorCode errorCode) {
-    return new ErrorResponse(errorCode.getErrorCode(), errorCode.getErrorMessage());
+    return new ErrorResponse(
+        errorCode.getErrorCode(),
+        errorCode.getErrorMessage()
+    );
   }
+
 }
